Move CrySystem project to CMake
- Add CMakeLists file and setup project
- Make necessary modifications in code
- Resource setup is not converted for now

+++ CrySystem/CryWaterMark.h
@@ -6,7 +6,7 @@
 // anywhere in the global scope in the program
 
 #define NUMMARKWORDS 10
-#define WATERMARKDATA(name) int name[] = { 0xDEBEFECA, 0xFABECEDA, 0xADABAFBE, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
+#define WATERMARKDATA(name) int name[] = { static_cast<int>(0xDEBEFECA), static_cast<int>(0xFABECEDA), static_cast<int>(0xADABAFBE), 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 };
 
 // (the name is such that you can have multiple watermarks in one exe, don't use
 // names like "watermark" just incase you accidentally give out an exe with
